Prevents rounding of the playback speed to set
Otherwise 0.5x will stop the video and everything below 2x doesnt change a thing.

Fixes: OX-8672

+++ lib/cinematic.ts
@@ -557,7 +557,7 @@ class Cinematic {
         if (!newSpeed) {
             return;
         }
-        this.speed = parseInt(newSpeed);
+        this.speed = parseFloat(newSpeed);
         this._video.playbackRate = this.speed;
     }
